refactor portfolio components into organized structure
