+++ frontend/src/app/portfolio/page.tsx
@@ -1,12 +1,4 @@
-// app/portfolio/page.tsx
-'use client';
-
-import dynamic from 'next/dynamic';
-
-// Dynamically import to avoid hydration mismatches or for better performance
-const PortfolioPage = dynamic(() => import('@/components/Portfolio/PortfolioPage'), {
-  ssr: false,
-});
+import PortfolioPage from "@/components/portfolio/PortfolioPage";
 
 export default function Portfolio() {
   return <PortfolioPage />;
