+++ frontend/src/components/portfolio/lib/currency.ts
@@ -1,0 +1,3 @@
+export function fmtCurrency(n: number) {
+  return n.toLocaleString(undefined, { style: "currency", currency: "USD", maximumFractionDigits: 0 });
+}
