+++ frontend/src/components/portfolio/lib/index.ts
@@ -1,0 +1,1 @@
+export { fmtCurrency } from "./currency";
