+++ frontend/src/components/portfolio/shared/AccountBalanceGraph.tsx
