+++ frontend/src/components/portfolio/shared/GainLossBarChart.tsx
