+++ frontend/src/components/portfolio/shared/PortfolioSummaryCard.tsx
