+++ frontend/src/components/portfolio/shared/PositionTable.tsx
