+++ frontend/src/components/portfolio/shared/TradingHistoryTable.tsx
