+++ frontend/src/components/portfolio/shared/TransactionsTable.tsx
