+++ frontend/src/hooks/usePortfolioData.ts
@@ -1,4 +1,4 @@
-// src/components/Portfolio/usePortfolioData.ts
+// src/hooks/usePortfolioData.ts
 import { useQuery } from '@tanstack/react-query';
 import { getActiveApiPortfolioData } from '@/api/brokerService';
